+++ src/Form/MonCompteType.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\User;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class MonCompteType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('prenom', TextType::class)
+            ->add('nom', TextType::class)
+            ->add('telephone', TextType::class, [
+                'required' => false,
+            ])
+            ->add('submit', SubmitType::class, [
+                'label' => 'Mettre à jour',
+                'attr' => ['class' => 'btn btn-success'],
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'data_class' => User::class,
+        ]);
+    }
+}